+++ src/TrelloDotNet/Model/Options/GetBoardOptions/GetBoardOptionsFilter.cs
@@ -1,5 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 
+// ReSharper disable UnusedMember.Global
+
 namespace TrelloDotNet.Model.Options.GetBoardOptions
 {
     /// <summary>